Refactor some code into components, and fix some build problems.
