+++ src/shared-components/character-upgrades.tsx
@@ -15,6 +15,7 @@ import { UpgradeControl } from 'src/shared-components/upgrade-control';
 import { StatCalculatorService } from 'src/v2/functions/stat-calculator-service';
 import { StaticDataService } from 'src/services';
 import { ICharacter2 } from 'src/models/interfaces';
+import { IUnit } from 'src/v2/features/characters/characters.models';
 
 interface Props {
     character: ICharacter2;
